Begins to implement event

+++ src/main/java/ch/roester/product_size_price/ProductSizePrice.java
@@ -3,7 +3,6 @@ package ch.roester.product_size_price;
 import ch.roester.product.Product;
 import ch.roester.unit.Unit;
 import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -20,17 +19,17 @@ public class ProductSizePrice {
 
 
     @ManyToOne
-    @JoinColumn(name="product_id", nullable=false)
+    @JoinColumn(name = "product_id", nullable = false)
     private Product product;
 
     @ManyToOne
-    @JoinColumn(name="unit_id", nullable=false)
+    @JoinColumn(name = "unit_id", nullable = false)
     private Unit unit;
 
     @Column(name = "amount", nullable = false)
-    private float amount;
+    private Double amount;
 
     @Column(name = "price", nullable = false)
-    private float price;
+    private Double price;
 
 }
